Refactor event model and model structure
