+++ backend/samfundet/models/general.py
@@ -1,15 +1,17 @@
+#
+# This file contains most of the django models used for Samf4
+#
+
 from __future__ import annotations
 
 import random
 import re
-import uuid
 from datetime import time, timedelta
 from typing import TYPE_CHECKING
 
 from django.contrib.auth.models import AbstractUser
 from django.core.exceptions import ValidationError
 from django.db import models
-from django.utils import timezone
 from django.utils.translation import gettext as _
 from guardian.shortcuts import assign_perm
 
@@ -88,169 +90,6 @@ class User(AbstractUser):
         has_global_perm = super().has_perm(perm=perm)
         has_object_perm = super().has_perm(perm=perm, obj=obj)
         return has_global_perm or has_object_perm
-
-
-class EventGroup(models.Model):
-    """
-    Used for recurring events
-    Connects multiple recurring events (e.g. the same concert two days in a row)
-    Enables frontend to know about recurring events and provide tools
-    for admins to edit both or links for users to see other times.
-    """
-    name = models.CharField(max_length=140)
-    created_at = models.DateTimeField(null=True, blank=True, auto_now_add=True)
-    updated_at = models.DateTimeField(null=True, blank=True, auto_now=True)
-
-    class Meta:
-        verbose_name = 'EventGroup'
-        verbose_name_plural = 'EventGroups'
-
-
-class NonMemberEmailRegistration(models.Model):
-    """
-    Non-member registration for events
-    Uses a unique id to enable cancellation of registration
-    """
-    # Long unique identifier (used for email cancellation)
-    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
-    # Email of the registered user
-    # WARNING: sensitive data! Make sure this is not exposed
-    # through the API or otherwise
-    email = models.EmailField(blank=False, null=False, editable=False)
-
-
-# Data for registration ticket type
-class EventRegistration(models.Model):
-    """
-    Used for events with registration payment type
-    Stores list of registered users and emails.
-    """
-    # Registered users
-    registered_users = models.ManyToManyField(User, blank=True, null=True)
-    # Registered emails (for those not logged in/not a member)
-    registered_emails = models.ManyToManyField(NonMemberEmailRegistration, blank=True, null=True)
-
-    @property
-    def count(self) -> int:
-        return self.registered_users.count() + self.registered_emails.count()
-
-
-# Custom ticket for "custom" event ticket type
-class EventCustomTicket(models.Model):
-    """
-    Used for events with custom price group.
-    Stores name and price of each custom ticket type.
-    """
-    name_nb = models.CharField(max_length=140, blank=False, null=False)
-    name_en = models.CharField(max_length=140, blank=False, null=False)
-    price = models.PositiveIntegerField(blank=False, null=False)
-
-
-class Event(models.Model):
-    # General event info/metadata
-    title_nb = models.CharField(max_length=140, blank=False, null=False)
-    title_en = models.CharField(max_length=140, blank=False, null=False)
-    description_long_nb = models.TextField(blank=False, null=False)
-    description_long_en = models.TextField(blank=False, null=False)
-    description_short_nb = models.TextField(blank=False, null=False)
-    description_short_en = models.TextField(blank=False, null=False)
-    location = models.CharField(max_length=140, blank=False, null=False)
-    image = models.ForeignKey(Image, on_delete=models.PROTECT, blank=False, null=False)
-    host = models.CharField(max_length=140, blank=False, null=False)
-
-    # Event group is used for events occurring multiple times (e.g. a concert repeating twice)
-    event_group = models.ForeignKey(EventGroup, on_delete=models.PROTECT, blank=True, null=True)
-
-    # Timestamps
-    created_at = models.DateTimeField(null=True, blank=True, auto_now_add=True)
-    updated_at = models.DateTimeField(null=True, blank=True, auto_now=True)
-    start_dt = models.DateTimeField(blank=False, null=False)
-    duration = models.PositiveIntegerField(blank=False, null=False)
-    publish_dt = models.DateTimeField(blank=False, null=False)
-
-    # Age restriction
-    class AgeRestriction(models.TextChoices):
-        NO_RESTRICTION = 'none', _('Ingen aldersgrense')
-        AGE_18 = 'eighteen', _('18 år')
-        AGE_20 = 'twenty', _('20 år')
-        MIXED = 'mixed', _('18 år (student), 20 år (ikke-student)')
-
-    age_restriction = models.CharField(max_length=30, choices=AgeRestriction.choices, blank=False, null=False)
-
-    # Event status
-    class Status(models.TextChoices):
-        """
-        Status for a given event. Deleted status
-        is used to hide event without actually deleting it
-        so that it can be restored if something wrong happens
-        """
-        ACTIVE = 'active', _('Aktiv')
-        ARCHIVED = 'archived', _('Arkivert')
-        CANCELED = 'cancelled', _('Avlyst')
-        DELETED = 'deleted', _('Slettet')
-
-    status = models.CharField(max_length=30, choices=Status.choices, blank=False, null=False, default=Status.ACTIVE)
-
-    # Event category
-    class Category(models.TextChoices):
-        """
-        Used for sorting, filtering and organizing stuff in frontend
-        """
-        SAMFUNDET_MEETING = 'samfundsmote', _('Samfundsmøte')
-        CONCERT = 'concert', _('Konsert')
-        DEBATE = 'debate', _('Debatt')
-        QUIZ = 'quiz', _('Quiz')
-        LECTURE = 'lecture', _('Kurs')
-        OTHER = 'other', _('Annet')
-
-    category = models.CharField(max_length=30, choices=Category.choices, blank=False, null=False, default=Category.OTHER)
-
-    # Price groups
-    class TicketType(models.TextChoices):
-        """
-        Handles event ticket type.
-            Included/Free - simple info shown on event
-            Billig - event connected to billig payment system
-            Registration - connect event to registration model (påmelding)
-            Custom - connect event to custom payment list (only used to show in frontend)
-        """
-        INCLUDED = 'included', _('Included with entrance')
-        FREE = 'free', _('Free')
-        BILLIG = 'billig', _('Paid')
-        REGISTRATION = 'registration', _('Free with registration')
-        CUSTOM = 'custom', _('Custom')
-
-    ticket_type = models.CharField(max_length=30, choices=TicketType.choices, blank=False, null=False, default=TicketType.FREE)
-    capacity = models.PositiveIntegerField(blank=False, null=False)
-
-    # Registration data for events with registration price group
-    registration = models.ForeignKey(EventRegistration, blank=True, null=True, on_delete=models.PROTECT, editable=False)
-
-    # Custom tickets for events with custom price group
-    custom_tickets = models.ManyToManyField(EventCustomTicket, blank=True, null=True)
-
-    # TODO billig event foreign key handling
-
-    @property
-    def total_registrations(self) -> int:
-        """
-        Total number of registrations made for registration type events.
-        """
-        if self.ticket_type == Event.TicketType.REGISTRATION and self.registration:
-            return self.registration.count
-        return 0
-
-    @property
-    def image_url(self) -> str:
-        return self.image.image.url
-
-    @property
-    def end_dt(self) -> timezone.datetime:
-        return self.start_dt + timezone.timedelta(minutes=self.duration)
-
-    class Meta:
-        verbose_name = 'Event'
-        verbose_name_plural = 'Events'
 
 
 class Venue(models.Model):
