+++ backend/samfundet/utils.py
@@ -1,9 +1,10 @@
 from __future__ import annotations
-from django.http import QueryDict
+
 from django.db.models import Q
 from django.db.models.query import QuerySet
+from django.http import QueryDict
 
-from .models import (
+from .models.event import (
     Event,
 )
